+++ anybox/recipe/openerp/tests/test_server.py
@@ -240,6 +240,10 @@ class TestServer(RecipeTestCase):
             if not script in binlist:
                 self.fail("Script %r missing in bin directory." % script)
 
+    def read_script(self, script_name):
+        with open(os.path.join(self.buildout_dir, 'bin', script_name)) as f:
+            return f.read()
+
     def test_retrieve_fixup_addons_local_61(self):
         addons_dir = os.path.join(self.buildout_dir, 'addons-custom')
         oerp_dir = os.path.join(TEST_DIR, 'oerp61')
@@ -403,6 +407,25 @@ class TestServer(RecipeTestCase):
 
     def test_install_scripts_70_no_devtools(self):
         self.test_install_scripts_70(with_devtools=False)
+
+    def test_install_scripts_70_server_wide_modules(self):
+        self.make_recipe(version='local %s' % os.path.join(TEST_DIR, 'oerp70'),
+                         gunicorn='direct',
+                         server_wide_modules='anybox_homepage')
+        self.recipe.version_detected = "7.0alpha"
+        self.recipe.options['options.log_handler'] = ":INFO,werkzeug:WARNING"
+
+        self.install_scripts(extra_develop={
+            'openerp-command': 'fake_openerp-command'})
+
+        self.assertTrue("server_wide_modules=('anybox_homepage',)"
+                        in self.read_script('start_openerp'))
+        with open(os.path.join(
+                self.buildout_dir, 'etc', 'gunicorn_openerp.conf.py')) as gu:
+            self.assertTrue(
+                "openerp.conf.server_wide_modules = "
+                "['web', 'anybox_homepage']\n"
+                in gu)
 
     def test_install_scripts_80(self, with_devtools=True, **kw):
         kw['with_devtools'] = 'true' if with_devtools else 'false'